Max tx power + gain <= max eirp The current version has different values for gain depending on if capacity or interference is being calculated. This can create inconsistencies.
Technically, the tx power can be increased if there is scan loss.
This would be an orientation-dependent max tx power. For now,
we opt for a orientation-independent max tx power that simply
takes the max gain (user supplied antenna boresight gain +
diversity gain) and subtracts that from max eirp to bound
the max tx power.

Making max tx power orientation-dependent can be a follow-up
to this fix.