CNs Tilt (#22)
CNs are mechanically tilted toward their serving DN. Update capacity calculation so that this is properly reflected (i.e., elevation deviation is 0 in this case).

Note: interference calculations are not impacted. Post-optimization, the angle is between links and sector orientation is not relevant. Pre-optimization, the interfering path deviations are used (a comment is added/updated for this).

Note: the capacity calculation for inactive CNs after optimization is slightly incorrect. It assumes 0 deviation although the CN is oriented toward a different DN. Because capacity calculations for inactive links is not that important, this issue is ignored.